fixed bug

+++ src/hstring.h
@@ -16,9 +16,10 @@
 #define DELIM_NOT_INIT  42
 
 /* 
- * Symbols for words.
+ * Symbols for words. Note: Some measures enumerate all possible symbols.
+ * These need to be patched first to support larger symbol sizes.
  */
-typedef uint64_t sym_t;
+typedef uint16_t sym_t;
 
 /* Support types for strings. See union str in struct */
 #define TYPE_CHAR		0x00
